二回目チェック用

+++ app/View/Books/detail.ctp
@@ -25,7 +25,7 @@
       <!--そのまま帰ってきたのだと前半にいいものが入っていない→そこから目的の本情報が入っている['Items']にまで言った。foreachの回し方に注意！-->
         
     <tr>
-        <td><?php echo $this->Html->link($detail['title'], 'detail?title='.$detail['title'], array('action' => 'detail')); ?></td><!--
+        <td><?php echo $this->Html->link($detail['title'], 'detail?isbn='.$detail['isbn'], array('action' => 'detail')); ?></td><!--
         詳細画面に行く時には、どのページの情報を取得すればいいかに注意。ここで、タイトルをクリックすると、パラメーターにタイトルを持った、books/detail/['title']に遷移。
         この時にしっかりとこのページでゲットした各種必要な変数をゲットする。また、isbnでしっかりと基さんが撮っていた本のあらすじを取れるようにするここから、books/detail?title=???というページに移行させる。ここではお気に入りボタンの表示、記事の詳細情報(検索結果の他に出版社、あらすじ、価格がある)
         ここで、その記事に紐付いているレビューを持ってくる。また、レビューページの遷移にもこのページの情報というか、そういうものが必要である。
@@ -41,21 +41,67 @@
     </tr>
     
 
-
-            
-    <h2>本の登録</h2>
-    <?php
-    echo $this->Form->create('Favorite', array('action' => 'add'));
-    echo $this->Form->input('title', array('type'=>'hidden','value'=>$detail['title']));
-    echo $this->Form->input('author', array('type'=>'hidden','value'=>$detail['author']));
-    echo $this->Form->input('publisher', array('type'=>'hidden','value'=>$detail['publisherName']));
-    echo $this->Form->input('text', array('type'=>'hidden','value'=>$detail['itemCaption']));
-    echo $this->Form->input('publication', array('type'=>'hidden','value'=>$detail['salesDate']));
-    echo $this->Form->input('price', array('type'=>'hidden','value'=>$detail['itemPrice']));
-    echo $this->Form->input('isbn', array('type'=>'hidden','value'=>$detail['isbn']));
-    echo $this->Form->input('thumbnail', array('type'=>'hidden','value'=>$detail['smallImageUrl']));
-    echo $this->Form->end(array('label' => '登録', 'name' => 'book'));
+    <?php 
+    if(empty($confirm)) {
+        echo '<h2>本の登録</h2>';
+        echo $this->Form->create('Favorite', array('action' => 'add'));
+        echo $this->Form->input('title', array('type'=>'hidden','value'=>$detail['title']));
+        echo $this->Form->input('author', array('type'=>'hidden','value'=>$detail['author']));
+        echo $this->Form->input('publisher', array('type'=>'hidden','value'=>$detail['publisherName']));
+        echo $this->Form->input('text', array('type'=>'hidden','value'=>$detail['itemCaption']));
+        echo $this->Form->input('publication', array('type'=>'hidden','value'=>$detail['salesDate']));
+        echo $this->Form->input('price', array('type'=>'hidden','value'=>$detail['itemPrice']));
+        echo $this->Form->input('isbn', array('type'=>'hidden','value'=>$detail['isbn']));
+        echo $this->Form->input('thumbnail', array('type'=>'hidden','value'=>$detail['smallImageUrl']));
+        echo $this->Form->end(array('label' => '登録', 'name' => 'book'));
+    } else {
+        echo '<h2>登録内容の修正</h2>';
+        echo $this->Html->link('編集', array('controller' => 'favorites', 'action'=>'edit', $confirm['Favorite']['id']));
+/*
+        echo '<h2>登録内容の修正</h2>';
+        echo $this->Form->create('Favorite', array('action' => 'edit'));
+        echo $this->Form->end(array('label' => '修正', 'name' => 'book'));
+        */
+    }
     ?>
     
     <?php endforeach; ?>
+</table>
+
+<h2>レビュー</h2>
+
+<?php  if(!empty($reviews)){
+    echo        '<table>';
+     
+    echo '<tr>';
+    echo    "<th>レビュワー名</th>";
+    echo        "<th>評価</th>";
+    echo   "<th>本文</th>";
+    echo  "</tr>";
+    echo    "<tr>";
+    
+    foreach ($reviews as $review):
+    
+    echo "<tr>";
+    echo  "<td>";
+    echo $this->Html->link($review['User']['username'], '/users/mypage/'.$review['User']['id']);
+    echo "</td>";
+    echo  "<td>";
+    echo $review['Favorite']['evaluation'];
+    echo "</td>";
+    echo  "<td>";
+    echo $review['Favorite']['review'];
+    echo "</td>";
+    echo "</tr>";
+    endforeach;
+    echo "</table>";
+    } else { 
+        echo '<h2>レビューはありません</h2>';
+    }
+?>
+
+     
+    
+    
+     
 </table>
